Moved main into a Main class, allowing Builder to be non-static
